% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xaf.R
\name{read.xaf}
\alias{read.xaf}
\title{Read Dutch XML Audit Files (XAF)}
\usage{
read.xaf(
  file,
  progress = interactive(),
  clean = TRUE,
  lang = c("nl", "en")
)
}
\arguments{
\item{file}{a string specifying the path to the XAF file to be read.}

\item{progress}{logical. Whether to show a progress bar.}

\item{clean}{logical. Whether to clean data before return.}

\item{lang}{character. Language of the added text and column names.}
}
\value{
A data frame containing the parsed data from the XAF file.
}
\description{
This function reads and parses Dutch XML Audit Files (XAF),
  which are commonly used in the Netherlands for financial and auditing
  purposes.
}
\examples{
\dontrun{
df <- read.xaf("path/to/xaf/file.xaf")
head(df)
attributes(df)$Journals
}
}
\references{
For more information about XAF files, see
  \url{http://www.auditfiles.nl}.
}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{read}
\keyword{xaf}
