% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaf_balance.R
\name{xaf_balance}
\alias{xaf_balance}
\title{Create a Balance Sheet from a cleaned XAF file}
\usage{
xaf_balance(x, date = NULL)
}
\arguments{
\item{x}{a data frame resulting from a call to \code{read_xaf()}.}

\item{date}{a character specifying the date up until the balance should be
made up. The required format is day-month-year (i.e., 02-01-2016 is january
second, 2016).}
}
\value{
A data frame containing the balance sheet.
}
\description{
This function creates a balance sheet from a cleaned XAF file
  at a specific point in time.
}
\examples{
\dontrun{
df <- read_xaf("path/to/xaf/file.xaf")
xaf_balance(df, date = "2016-03-08")
}
}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{balance}
\keyword{xaf}
