% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaf_balance_sheet.R
\name{xaf_balance_sheet}
\alias{xaf_balance_sheet}
\title{Create a Balance Sheet from a Cleaned XAF File}
\usage{
xaf_balance_sheet(x, date = NULL)
}
\arguments{
\item{x}{a data frame resulting from a call to \code{read_xaf()}.}

\item{date}{a character specifying the date up until the balance should be
made up. The required format is day-month-year (i.e., 02-01-2016 is january
second, 2016).}
}
\value{
A data frame containing the balance sheet.
}
\description{
This function creates a balance sheet at a specific point in
  time from a cleaned XAF file.
}
\examples{
\dontrun{
df <- read_xaf("path/to/xaf/file.xaf")
xaf_balance_sheet(df, date = "31-12-2022")
}
}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{balance}
\keyword{sheet}
\keyword{xaf}
