% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaf_subtables.R
\name{xaf_subtables}
\alias{xaf_subtables}
\title{Extract Subtables from a XAF file}
\usage{
xaf_subtables(x)
}
\arguments{
\item{x}{a data frame resulting from a call to \code{read_xaf()}.}
}
\value{
A list containing the journals, accounts, vat codes and relations in
  the XAF file.
}
\description{
This function extracts the journals and accounts subtables from
  a XAF file. If the XAF file contains additional tables such as the VAT
  Codes and Customer / Supplier data, these are extracted as well.
}
\examples{
\dontrun{
df <- read_xaf("path/to/xaf/file.xaf")
xaf_subtables(df)
}
}
\author{
Koen Derks, \email{k.derks@nyenrode.nl}
}
\keyword{subtables}
\keyword{xaf}
